refactor(codegen): extract std::crypto::* into stdlib/crypto.rs (Round 1.6)
4 fns lifted into CryptoStdlib<'ctx>: sha1, sha256, hmac_sha256, crc32.
No new pub(crate) surface needed.

codegen.rs: 43984 -> 43814 (-170 lines).
